[util] introduce ThreadPoolToken::Close()
This patch introduces a means to protect a thread pool token from
submission of new tasks, while allowing all of its queued tasks to
complete before the token reaches terminal QUIESCED state.  That's the
main difference from ThreadPoolToken::Shutdown(), where all the token's
queued but not-yet-in-flight tasks are thrown out of the thread pool's
queue.  From this perspective, the newly introduced Close() method might
be considered as a graceful and asynchronous version of its Shutdown()
counterpart.

This patch also adds a few test scenarios to cover the crucial parts
of the newly introduced functionality.

Change-Id: Ic226319c191af666233ec735c1787389d051b039
Reviewed-on: http://gerrit.cloudera.org:8080/22614
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>