[util] fix flakiness in threadpool-test
There were reports about a few test scenarios failing intermittently
with errors like below on a very busy machine:

  [ RUN      ] ThreadPoolTest.TestThreadPoolWithNoMinimum
  /root/Projects/kudu/src/kudu/util/threadpool-test.cc:317: Failure
  Expected equality of these values:
    0
    pool_->num_threads()
      Which is: 1

  [ RUN      ] ThreadPoolTest.TestVariableSizeThreadPool
  /root/Projects/kudu/src/kudu/util/threadpool-test.cc:459: Failure
  Expected equality of these values:
    1
    pool_->num_threads()
      Which is: 2

This patch fixes the issue, removing implicit assumptions on timings
of events that might not be true in the presence of scheduled anomalies.

This is a follow-up to fc8615c37eb4e28f3cc6bea0fcd5a8732451e883.

Change-Id: I8caed555b2207d25a93105037d77bd579d160550
Reviewed-on: http://gerrit.cloudera.org:8080/22622
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>