[util] use std::atomic in HighWaterMark
This patch replaces chromium-based Atomics with std::atomic in the
implementation of HighWaterMark class.  This is a part of the on-going
work to replace chromium-derived atomics with STL-based ones.

While TryIncrementBy() provides 'a strict limit' semantics as it should,
the rest of the read-style methods in HightWaterMark API adhere to using
the std::memory_order_relaxed ordering, similar to the original version.
That's because all the use cases of such methods involve TOCTOU
conditions anyway.  With that, it wouldn't be a good idea switching to
more strict ordering for read operations (i.e. std::memory_order_acquire
for load) and incur a performance hit without a clear benefit in return.

Change-Id: I17e4d4cf308c06b0dd135a99e793067980d5ad20
Reviewed-on: http://gerrit.cloudera.org:8080/22629
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
(cherry picked from commit b1a4f5b4deec357159ac99477ac3479c005c3df2)
Reviewed-on: http://gerrit.cloudera.org:8080/22639