Add SecurityManagerCompatibility shim
This is a shim to allow the use of both modern (JRE 18+)
and legacy APIs (deprecated in 17, degraded in 23, removal unknown)
related to JEP 411: Deprecate the Security Manager for Removal.

This shim implements the interim strategy - Remove legacy APIs
are used as long as they are available and functional.
Once the legacy APIs are removed or degraded, the modern APIs are used instead.

Change-Id: I2a0f7306d9471c6a86bff3e4864947a0603337d7
Reviewed-on: http://gerrit.cloudera.org:8080/22703
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>