[compaction] Add memory tracking for better OOM issues triaging
This patch adds memory accounting for RowSet compaction.
The memory usage is tracked under newly added MemTracker.
The tracker is added per rowset compaction operation. The same
tracker is shared across multiple compaction inputs that are
involved in that specific compaction operation. Each input
has its own usage added to shared memory tracker during the
initialisation and is deducted in corresponding destructor.
Along with that, throttled warning messages are added when the
memory usage goes beyond a certain threshold.

A unit test is added to verify the presence of warning messages
when memory threshold (reduced for test purpose) is crossed.

Change-Id: I700fdf5ee0fbf0f2826b863d23d599f03ad4ac49
Reviewed-on: http://gerrit.cloudera.org:8080/22708
Reviewed-by: Zoltan Martonka <zmartonka@cloudera.com>
Tested-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Marton Greber <greber.mx@gmail.com>