[integration-tests] fix result status handling
This patch addresses sloppy result handling of functions/methods
that return Status under the src/kudu/integration-tests directory.

After I added proper result status handling, I found that the
ScanYourWritesMultiClientsParamTest.Test scenario in consistency-itest
wasn't doing what it was supposed to do.  I added a TODO blurb to
be addressed later a separate changelist because it's not trivial.
My goal for this changelist is to fix result status handling, and
maybe address trivial underlying bugs and issues that might be lurking
there due to sloppy result status handling.  However, addressing
a non-trivial bug or rewriting a test should be taken care of
in a separate changelist.

Change-Id: I09f0266063d5e6b02c668237047cdea04c6911a7
Reviewed-on: http://gerrit.cloudera.org:8080/22730
Tested-by: Kudu Jenkins
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Zoltan Martonka <zmartonka@cloudera.com>