[util] optimise RLE a bit
Since the number of bits for an RLE encoder/decoder doesn't change
and is known beforehand, it makes sense to make it a template parameter,
allowing the compiler to use compile-time optimisations when possible.
There are just a few pre-defined numbers of bits used for RLE
encoding/decoding in the production code, so there isn't a risk of
code bloat.

Change-Id: I0951b7efc8f4a11ec89ccc0dda1be3faf4cc023c
Reviewed-on: http://gerrit.cloudera.org:8080/22884
Reviewed-by: Zoltan Martonka <zmartonka@cloudera.com>
Tested-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Tested-by: Alexey Serbin <alexey@apache.org>