KUDU-3663: Support certificates signed with RSASSA-PSS for channel bindings
The existing code to determine the hash algorithm for a certificate
does not handle RSASSA-PSS signatures as the hash algorithm is
configurable for RSASSA-PSS. OpenSSL 1.1.1 introduced the
x509_get_signature_info() function, which is able to determine the
hash algorithm even for RSASSA-PSS. This uses x509_get_signature_info()
whenever building against OpenSSL 1.1.1 or above. This is similar to
the fix used in Postgres when faced with the same issue.

Testing:
 - Added a test certificate that uses RSASSA-PSS and a test case in
   cert-test to verify that it can determine the hash algorithm used

Change-Id: I26a25a43d778fd2f2fcf293ecb199133c675212b
Reviewed-on: http://gerrit.cloudera.org:8080/22910
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>