KUDU-1261 introduce nested types for ColumnSchemaPB
This changelist introduces changes in the common protobuf structures
to allow for columns of nested data types in Kudu tables.  At the time
of writing, we are planning to support only one-dimensional arrays
of scalar types as a first phase, but it's possible to extend the newly
introduced protobuf messages to support structures, maps, and type
nesting as necessary: I added TODOs in the comments.

Follow-up patches will start using the auto-generated bindings,
but this changelist is focused on introducing protobuf-related
changes only.

Change-Id: Icf5f32c4291e7906ffb68070bee0d5369fab36fd
Reviewed-on: http://gerrit.cloudera.org:8080/22925
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>