KUDU-3698 fix NPE in TableLocationsCache.get() in kudu-client
It might happen that the special table location entry for Kudu master
is absent in the cache, but the code since [1] doesn't check for null
before calling isStale() on the item returned from the cache.  This
results in a NPE in Kudu Java client.

This patch addresses the issue.  This is a follow-up to [1].

[1] https://github.com/apache/kudu/commit/a50aa5bfd

Change-Id: Ic1bf9001eb88e85c2d9913b804d3f223ab022cdc
Reviewed-on: http://gerrit.cloudera.org:8080/22958
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit adc8f2f84e5b909a1e8bcb030a34139b21bda94c)
Reviewed-on: http://gerrit.cloudera.org:8080/23415