[util] Remove SchedulerThread code and tests
This removes the SchedulerThread implementation and its associated tests.
The TestSubmitAfterShutdown case has proven unstable in ARM+TSAN
environments. The code was originally introduced in KUDU-3364 but has
never been used outside of tests, and the original author appears to
have no plans to maintain it.

Removing this unused code simplifies the codebase and improves test
stability. The removal was suggested by Alexey.

Change-Id: I544488158db84dca580694e884b56d5c546d89c1
Reviewed-on: http://gerrit.cloudera.org:8080/22987
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>