KUDU-1261 support 1D arrays in TypeInfo
This changelist updates TypeInfo to support one-dimensional arrays
of Kudu scalar types.  With this patch, the run-time registry of data
types is now populated with necessary TypeInfo objects for 1D arrays
for every registered Kudu scalar data type.

This changelist also adds basic tests scenarios to verify the contents
of the TypeInfo registry once the necessary array types have been added.

The following methods aren't implemented yet as a part of this changelist
and will be implemented in a follow-up changelist after the introduction
of the ArrayCellMetadataView class on top of [1]:
  * ArrayTypeTraits::Compare()
  * ArrayTypeTraits::AreConsecutive()
  * ArrayTypeTraits::AppendDebugStringForValue()

[1] https://gerrit.cloudera.org/#/c/23002/

Change-Id: I4215a86c1989d5bc37a6980312fa7fc82f136d60
Reviewed-on: http://gerrit.cloudera.org:8080/23001
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>