[java] KUDU-3657: fix publishing
The changes introduced in KUDU-3610 have fixed the POM generation for
the kudu-hive java module, however the publishing for other modules were
broken as a side-effect. This patch aims to fix these issues.
The change in publishing.gradle file checks if the implementation and shadow
configurations are present to calculate the correct pom to be generated.
If they are not present for the given module, they are simply filled out
as empty arrays.

Additionally, the skipping of publishing wasn't working, this needed to
be fixed for the following modules (to keep in alignment with pre-gradle
upgrade behaviour):
* kudu-backup-common
* kudu-jepsen
* kudu-proto
* kudu-subprocess

Fixed the generated jar names of the following modules:
* kudu-backup -> kudu-backup3_2.12
* kudu-spark -> kudu-spark3_2.12
* kudu-spark-tools -> kudu-spark3-tools_2.12

Changed the compileUnshaded configuration to be extending the
compileOnly configuration. This was necessary to correct the scope of
these dependencies. This change also made it necessary to add the
compileUnshaded dependencies explicitly to the test compile path as
well[0].

[0] https://blog.gradle.org/introducing-compile-only-dependencies

Change-Id: Ie045805a68bfcd16325e1e4dbb38653cefe3a4b0
Reviewed-on: http://gerrit.cloudera.org:8080/22768
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit 6d72d4a242076468501f3430b9a2cd050c634be2)
Reviewed-on: http://gerrit.cloudera.org:8080/23079
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Tested-by: Marton Greber <greber.mx@gmail.com>