[docs] refresh list of supported OSes and distros
This patch adds information on contemporary OSes and distros where
Kudu builds and runs for the head of the 'master' and 'branch-1.18.x'
git repo branches.

I didn't remove Linux and macOS distributions that are EOL at the time
of writing: Kudu still builds and runs at those.  Besides, some of the
EOL Linux distros are still in the 'extended support' lifecycle.

I also changed 'C++' --> '{cpp}' for proper adoc-to-html rendering.

Change-Id: I5a58460381a58ced6f3b90a156ad62be5512e579
Reviewed-on: http://gerrit.cloudera.org:8080/23014
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>