[thirdparty] don't use system curl for cmake build
An update was introduced with curl v8.13.0 [1] that breaks cmake 3.25.3
compilation if using curl's v8.13.0 and newer header files.

This changelist disables using system curl header files for a while.
Hopefully, next cmake version in Kudu's 3rd-party will have the issue
addressed, so it will be possible to use system curl again to speed up
3rd-party compilation.

It would be possible to make a decision on using the system curl based
on the version of libcurl reported by pkg-config, but I'm not exactly
sure it's worth it.

[1] https://github.com/curl/curl/commit/2ec00372a

Change-Id: I2ec9f079667a374f04664abf4a9e840c86b3b7bf
Reviewed-on: http://gerrit.cloudera.org:8080/23048
Tested-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Gabriella Lotz <lotzgabriella@gmail.com>
Reviewed-by: Marton Greber <greber.mx@gmail.com>