SPNEGO authentication in ExternalMiniCluster
This commit edits the CreateKerberosConfig method to make calls to the
webserver work with SPNEGO authentication in the ExternalMiniCluster.
The changes include:

1. Creating a single keytab file that contains both the RPC service
   principal (kudu/host) and the HTTP service principal (HTTP/host).

2. Enabling SPNEGO authentication for the web UI by adding the
   appropriate flag (--webserver_require_spnego).

This change is necessary to properly test REST API authentication in
master_authz-itests.cc, as HTTP requests require SPNEGO to be configured
correctly to authenticate through Kerberos.

Some modifications, like setting authentication type to SPNEGO when
using EasyCurl in kerberized environments, had to be done in tests in
order for them to work as expected.

Change-Id: Ib6035bdf7f2f0f08f53485a7a1aa82ed26148cc4
Reviewed-on: http://gerrit.cloudera.org:8080/23052
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>