[compile] Fix compiler warnings about -Wrange-loop-construct
This commit resolves multiple instances of the
`-Wrange-loop-construct` warning across various tests.
These warnings occur when using `const std::string&` as
a loop variable to iterate over a brace-enclosed initializer
list of string literals (`const char*`).

The warnings were of the form:

    warning: loop variable '...' of type 'const string&' binds
    to a temporary constructed from type 'const char* const'
    [-Wrange-loop-construct]

To address this:

- Use of `std::string_view`.
- Use of `const std::string` directly in loop.
- Support for `std::string_view` in `SubstituteArg`.

This change eliminates warnings while maintaining safe and
readable code, and avoids unnecessary string copies when not
needed.

Change-Id: I32a9a043e7aade2c2c879472c13cbf2641daa7f3
Reviewed-on: http://gerrit.cloudera.org:8080/23057
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>