KUDU-3671 severity level filtering for /metrics_prometheus
This changelist implements severity level filtering for the
/metrics_prometheus endpoint of the embedded Kudu webserver based
on the current setting for the --metrics_default_level flag.  This
is now similar to the filtering for the /metrics endpoints, but the
severity level filtering for the /metrics_prometheus cannot be
overridden by query options as of this changelist.  Basic test
scenarios for Kudu master and tablet server are included.

Change-Id: If1542476f06d575594e8b3636cab3681f1b6fdfa
Reviewed-on: http://gerrit.cloudera.org:8080/23093
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit f7c956859e2f49c4cf1caffa969c1777a7a5d81c)
Reviewed-on: http://gerrit.cloudera.org:8080/23095
Tested-by: Abhishek Chennaka <achennaka@cloudera.com>