[tests] fix flakiness in exactly_once_writes-itest
This changelist fixes a flakiness in the
TestWritesWithExactlyOnceSemanticsWithCrashyNodes scenario of the
ExactlyOnceSemanticsITest test by postponing the fault injection until
after the start of the test cluster.

As one can see from the log snippet below, an unexpected crash happened
in TabletServerIntegrationTestBase::BuildAndStart() because of the crash
induced by the --fault_crash_after_leader_request_fraction flag's
setting.  Apparently, it wasn't intended to crash in BuildAndStart().

I also added the 'runtime' tag for a couple of test-only flags defined
in src/kudu/consensus/consensus_peers.cc and re-factored the relevant
code in exactly_once_writes-itest.cc to address clang-tidy warnings.

  E20250626 08:41:09.165330 27745 fault_injection.cc:59] injecting fault for kudu (pid 27730): FLAGS_fault_crash_after_leader_request_fraction (process will exit)
  W20250626 08:41:09.188339 27092 connection.cc:537] server connection from 127.26.53.135:35457 recv error: Network error: recv error from unknown peer: Transport endpoint is not connected (error 107)
    ...
  src/kudu/integration-tests/ts_itest-base.cc:555: Failure
  Failed
  Bad status: Network error: Client connection negotiation failed: client connection to 127.26.53.135:44789: BlockingRecv error: recv error from unknown peer: Transport endpoint is not connected (error 107)
  src/kudu/integration-tests/exactly_once_writes-itest.cc:208: Failure
  Expected: BuildAndStart(ts_flags, master_flags) doesn't generate new fatal failures in the current thread.
    Actual: it does.

Change-Id: Ia2d9da084220fa48749fa67cfb4dca422a9601f2
Reviewed-on: http://gerrit.cloudera.org:8080/23099
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>