[util] clean up on SCOPED_OPENSSL_NO_PENDING_ERRORS
Add 'const' for ScopedCheckNoPendingSSLErrors instances in the
SCOPED_OPENSSL_NO_PENDING_ERRORS macro and instantiate those only
if DCHECK_IS_ON() since the corresponding constraints are enforced
with the DCHECK_EQ() macro.

The motivation for this change was seeing warnings in QtCreator IDE
on improving const-correctness for ScopedCheckNoPendingSSLErrors
instances as part of the SCOPED_OPENSSL_NO_PENDING_ERRORS macro.

This changelist doesn't contains any functional modifications.

Change-Id: I94a9b6d34b36567b414599fdd10db64c932074d4
Reviewed-on: http://gerrit.cloudera.org:8080/23148
Reviewed-by: KeDeng <kdeng.ke@gmail.com>
Tested-by: KeDeng <kdeng.ke@gmail.com>