KUDU-3678: Fix crcutil building on MacOS arm64
When building crcutil, it uses 'uname -p' to determine the
processor, which prints 'arm'. This means that it doesn't match
against the correct processor and adds -msse2 -msse4.2 to
crcutil_flags, neither of which are supported on Apple silicon. This
adds a new branch that specifically targets MacOS devices running on
arm, allowing crcutil to compile correctly.

Change-Id: I71a1e70a8bc29732a8d0b31724bdbf8e00e7a736
Reviewed-on: http://gerrit.cloudera.org:8080/23198
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>