[net] use AF_INET hint for GetFQDN()
I noticed that on nodes with not properly configured IPv6 DNS resolver,
GetFQDN() would take quite a long time (~5 seconds) even if it
eventually returns the name.  This patch addresses the issue, adding
the AF_INET hint for 'ai_family' when invoking getaddrinfo().  Doing so
is in sync with the set of hints provided for the invocation of
getaddrinfo() in the context of HostPort::ResolveAddresses().

Change-Id: Ibca0a3c1f7ad03a6f847057c38c00b42275dec8b
Reviewed-on: http://gerrit.cloudera.org:8080/23213
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Gabriella Lotz <lotzgabriella@gmail.com>
Reviewed-by: Ashwani Raina <araina.git@gmail.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
(cherry picked from commit 9c703fa5bd476c17441361de68a518847d23e001)
Reviewed-on: http://gerrit.cloudera.org:8080/23215
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Tested-by: Marton Greber <greber.mx@gmail.com>