[util] explicit relationship between Thread and ThreadMgr
This changelist makes the dependency between Thread and ThreadMgr
instances explicit.  It makes the code more robust, so it now works
even for edge cases of a program's start and shutdown.

This patch is a building block for a few more follow-up patches that
address the issue of kudu CLI tool crashing at shutdown sometimes.

Change-Id: Id3df2b617a5f648d375e5fb6c6b08e8454e72784
Reviewed-on: http://gerrit.cloudera.org:8080/23260
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>