KUDU-2439 address the issue up to some extent
This changelist adds a work-around for the issue described in KUDU-2439.
It's based on the recently added provisions for calling OPENSSL_cleanup
explicitly with the library of versions 1.1.1 and newer.
Yes, it's a stop-gap, but it's better than nothing.

As for testing, I verified that the issue is gone after seeing its
manifestation with the frequency of about 1 in 30 runs of the kudu CLI
tool with RELEASE bits on RHEL8.8 x86_64 and RHEL9.2 x86_64 without
the patch.  I ran the following for reproduction, changing 100rep
to 1000rep for verificaion runs:

  ./bin/kudu-tool-test --gtest_repeat=100

Without the patch, I saw core files left by the kudu CLI binary during
every 100rep run, where many of the core files would have stack traces
attributable to the JIRA item.  With this patch, not a single crash
has been observed and no core files have been generated after many
100rep and a few 1000rep test runs with Kudu RELEASE bits.

NOTE: the issue no longer manifests itself in RELEASE builds,
      but there is still something that makes kudu CLI sometimes crash
      on exit in DEBUG builds, but the stack trace doesn't involve any
      OpenSSL symbols: instead, it has mostly libtcmalloc symbols
      and also involves __do_global_dtors_aux() from librocksdb

Change-Id: I472bb6b5a4edf2a1a03c0e3cdcd64e743b7e1e1f
Reviewed-on: http://gerrit.cloudera.org:8080/23262
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>