KUDU-3687 higher timeout for WAL segment count in heavy workloads
As it turns out, even with the increased default timeout parameter for
WaitForMinFilesInTabletWalDirOnTS() in [1], a couple of new tests added
after that [2] are failing from time to time under TSAN due to heavy
workloads.  This changelist addresses the issue, increasing the timeout
argument for WaitForMinFilesInTabletWalDirOnTS() in a couple of
particular call sites from 60 up to 120 seconds.

[1] http://gerrit.cloudera.org:8080/1793
[2] http://gerrit.cloudera.org:8080/16726

Change-Id: I68f30bc8b1c59843b1485bd0924ddd51d144eff1
Reviewed-on: http://gerrit.cloudera.org:8080/23286
Reviewed-by: Gabriella Lotz <lotzgabriella@gmail.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>