KUDU-1261 add array.fbs into common/serdes
One of the previous changelists added Flatbuffers into the Kudu's
3rd-party, along with basic and benchmark tests.

This changelist adds a separate Flatbuffers IDL file array.fbs for
serdes-ing of Kudu 1D array types into the common/serdes directory.
The idea is to keep the earlier introduced (and almost identical as
of now) .fbs IDL file just for benchmarking, tests, and experiments,
so these two might evolve independently.

This is necessary for follow-up changelists.

Change-Id: Ie2c25aa32e3a72aada61e8980eb2613f8f649481
Reviewed-on: http://gerrit.cloudera.org:8080/23364
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>