[build] detect undefined but evaluated macros
Undefined macros that are used in pre-processor expressions might lead
to a confusion and unexpected results.  One recent example of that
was [1] where in PS6 the OPENSSL_VERSION_NUMBER macro was not defined
and silently evaluated to 0 in webserver_options.cc.  The issue was
mitigated by including the necessary header file in PS7.

With this update, both GCC and CLANG exit with an error if there is
a non-defined, but evaluated macro in the code.  I verified it worked
as expected with the described issue in [1] as of PS6.  Also, a few
updates were necessary to the already existing code to compile the code
without an error once the guard against such conditions was enabled.
Where necessary, I silenced the corresponding warning since it was
mostly the code imported into Kudu from other projects.

[1] https://gerrit.cloudera.org/#/c/23295

Change-Id: I9f60f08b41c675a2a7307664a2ef0c7ecf8d682e
Reviewed-on: http://gerrit.cloudera.org:8080/23367
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>