KUDU-3697: Bundle kudu-spark classes
Update the kudu-spark dependency configuration to implementation to
ensure its classes are bundled into the kudu-backup and
kudu-spark-tools JARs.

Manually tweak the Gradle publishing process so that this dependency
is declared as a compile-time dependency. There are likely better ways
to address this, hence the TODO.

This restores the previous behavior from before the Gradle upgrade:
bundling kudu-spark into the subprojects while keeping the dependency
scope as compile.

Change-Id: I8d3e3c95efe161adc916332477f8cd9f35c48e38
Reviewed-on: http://gerrit.cloudera.org:8080/23407
Tested-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Reviewed-by: Marton Greber <greber.mx@gmail.com>