KUDU-3660 do not run AdminCliTest.TestRebuildTables under TSAN
The AdminCliTest.TestRebuildTables often (almost certainly) fails when
running under TSAN with --stress_cpu_threads=N, where N is the number
of CPU cores on the test machine.  With the increase of the number
of retries and timeout intervals it might require several minutes to
complete if running with --stress_cpu_threads=N.  With that, I think
the most appropriate resolution is simply not run it under TSAN:
we do have enough coverage for the functionality with RELEASE, DEBUG,
and ASAN builds, and the building blocks of the scenario are covered
in TSAN builds in many other test scenarios.

In addition, this changelist adds SKIP_IF_SLOW_NOT_ALLOWED() into the
scenario: in my experiments, the fastest run of RELEASE bits of this
scenario was about 5 seconds on 8 core 2GHz CPU machine without
--stress_cpu_threads, while the informal threshold for slow Kudu tests
is 3 seconds.

Change-Id: Ia09bba0cb6f014ffd46529d4fec607e19caf1d2c
Reviewed-on: http://gerrit.cloudera.org:8080/23414
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>