[thirdparty] upgrade cpplint up to 1.6.1 release
The prior version of the cpplint tool from the google-styleguide project
[1] was from 2016 and it wasn't aware of the C++17 constructs such as
'if constexpr (...) {...}', so it would error out on such a code with
messages like below:

  If/else bodies with multiple statements require braces  [readability/braces]

At the time of writing, the latest released version of the cpplint tool
is 2.0.2 [2].  However, version 2.0.0 dropped Python2 and Python3.7
support, and it seems the 2.0.x release line is not quite stable yet.
The safest bet at this time is 1.6.1 which includes the fix for the
issue [3] and is the latest release in the 1.x line.  It was released
about 3 years ago in August 2022.

Since the cpplint tool has been moved into a separate repo [4] and
it isn't a part of the google-styleguide project repo [1] anymore,
I renamed corresponding entities for this Kudu's 3rd-party component
and updated the scripts accordingly:

  GSG --> CPPLINT

I also updated the list of cpplint filters to make the code pass the
checks of the upgraded cpplint tool.  That should be addressed in
a separate follow-up changelist(s), and it might entail reconciliation
with the IWYU tool suggestions on the same set of header files, patching
the cpplint tool itself, addressing many of the runtime/string issues,
and adding NOLINT() where addressing the reported issues doesn't seem
viable:

  * runtime/string is disabled since it doesn't recognize static
    function declarations like in src/kudu/client/client.h, so it
    will require either fixing it in cpplint itself or adding
    NOLINT(runtime/string) as necessary

  * build/include_what_you_use is disabled because it suggests including
    the <utility> header in many places, but the IWYU tool itself
    insists on removing it instead

[1] https://github.com/google/styleguide
[2] https://github.com/cpplint/cpplint/releases/tag/2.0.2
[3] https://github.com/cpplint/cpplint/commit/6e3e178b0
[4] https://github.com/cpplint/cpplint

Change-Id: I390dad51e8e0a57e5823ad789f7f8b0029e5bdb4
Reviewed-on: http://gerrit.cloudera.org:8080/23447
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Gabriella Lotz <lotzgabriella@gmail.com>